python-app/requirements.txt
Flask==2.0.3
pyspark==3.2.0
kafka-python==2.0.2
numpy==1.21.2
pandas==1.3.3